app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
databases[postgresql]==0.8.0
asyncpg==0.29.0
pydantic==2.5.0
scikit-learn==1.3.2
numpy==1.24.4
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
redis==4.5.1          # <= IMPORTANT: must be <5.0.0 for fastapi-cache2
fastapi-cache2[redis]==0.2.1